+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 ## 1.22.7
+* Updated dependencies
 * Migrate now also supports semicolon.
 
 ### Before migrate: `package.json`