+++ CHANGELOG.md
@@ -1,4 +1,39 @@
 # Change Log
+## 1.22.7
+* Updated dependencies
+* Migrate now supports semicolon.
+* Migrate now supports the `cd` command.
+
+### Before migrate: `package.json`
+``` JSON
+{
+  "scripts": {
+    "serve:dev:uva": "echo serve && cd node_modules && echo dev && echo uva",
+    "serve:acc:uva": "echo serve ;  cd node_modules ;  echo acc ;  echo uva"
+  }
+}
+```
+
+### After migrate: `launcher-config.json`
+``` JSON
+{
+  "scripts": {
+    "serve:dev:uva": [
+      "echo serve",
+      "node_modules",
+      "echo dev",
+      "echo uva"
+    ],
+    "serve:acc:uva": [
+      "echo serve || true",
+      "node_modules",
+      "echo acc || true",
+      "echo uva"
+    ]
+  }
+}
+```
+
 ## 1.22.6
 ### Changes
 * Updated starter templates